Added customer dashboard to view public models by providing a new component PublicModels.jsx and updating CustomerDashboard.jsx to include it.
This provides the user with a way to browse public
assets without needing to log in or register.

this also bypasses the need for scraper that can
bypass api blockers and rate limits, as the public
models are fetched directly from the database.

users can now view and download public models
directly from the customer dashboard. and copy the
model link or download the model file to share
with the maker.

TO DO
fix printables query link
allow the Maker to change price and send the
quote to the customer. Allow the user to request
a print without the need to upload a stl file.